fix(inject): report the pad fields a real controller fills
Three report fields carried a constant where the hardware carries state. The
touch points' contact id was the slot index, so a tap, a lift and a second
tap looked like one contact that never ended to anything detecting taps by
id. The DualSense's 32-bit packet sequence stayed zero. A virtual Switch Pro
had no colours in SPI flash, which draws a black body with black buttons.

The contact id is now minted per landing from one counter across both
points — an id names a contact, not a slot, so two live contacts cannot
share one. The report counter widens to 32 bits and feeds both the byte at
struct offset 6 and the packet sequence at 11.

The Switch serial at `0x6000` is left alone: it would need to differ per pad
or two virtual controllers merge on it, as they do on a shared MAC.

+++ crates/pf-inject/src/inject/linux/dualsense.rs
@@ -62,7 +62,9 @@ impl DsUhidIdentity {
 /// Virtual DualSense on `/dev/uhid`. Drop sends `UHID_DESTROY` and unbinds `hid-playstation`.
 pub struct DualSensePad {
     fd: File,
-    seq: u8,
+    /// Report counter: the low byte is the report's `seq_number`, the whole of it the
+    /// packet sequence at struct offset 11.
+    seq: u32,
     clock: SensorClock,
 }
