+++ crates/pf-inject/src/inject/windows/dualsense_windows.rs
@@ -236,7 +236,9 @@ pub struct DsWinPad {
     _sw: Option<super::gamepad_raii::SwDevice>,
     channel: PadChannel,
     attach: super::gamepad_raii::DriverAttach,
-    seq: u8,
+    /// Report counter: the low byte is the report's `seq_number`, the whole of it the
+    /// packet sequence at struct offset 11.
+    seq: u32,
     clock: SensorClock,
     /// v2.3 input-seqlock generation — see [`publish_input`].
     input_gen: u32,
